Use variable substitution/dirname with 0 instead of cd&& $OLDPWD
It is cleaner and makes the first cd redundant.
As there's no cd into the $ANDROID_ROOT first we have to find it by
hand first. Because this project is to be run in an android repo tree
we can assume that the directory that contains /.repo/ is our $ANDROID_ROOT.

Signed-off-by: Björn Bidar <theodorstormgrade@gmail.com>